+++ b3j0f/conf/configurable/logger.py
@@ -29,6 +29,7 @@ __all__ = ['Logger']
 
 from logging import Formatter, getLogger, FileHandler, Filter
 
+from os import environ
 from os.path import join, sep
 
 from b3j0f.utils.property import addproperties
@@ -37,6 +38,27 @@ from ..model.cat import Category
 from ..model.param import Parameter
 
 from .core import Configurable
+
+from inspect import getargspec, getmembers, isfunction
+
+from functools import wraps
+
+B3J0F_LOG_PATH = 'B3J0F_LOG_PATH'  #: log path environment variable.
+
+LOG_PATH = '.'  #: default log path.
+
+# update log_path with environment variable b3j0f log path
+if B3J0F_LOG_PATH in environ:
+    LOG_PATH = environ[B3J0F_LOG_PATH]
+
+
+def _filehandler(configurable):
+    """Default logging file handler."""
+
+    filename = configurable.log_name.replace('.', sep)
+    path = join(configurable.log_path, '{0}.log'.format(filename))
+
+    return FileHandler(path, mode='a+')
 
 
 def _updatelogger(self, kwargsvalue, name):
@@ -47,7 +69,8 @@ def _updatelogger(self, kwargsvalue, name):
 @addproperties(
     names=[
         'log_debug_format', 'log_info_format', 'log_warning_format',
-        'log_error_format', 'log_critical_format', 'log_name', 'log_path'
+        'log_error_format', 'log_critical_format', 'log_name', 'log_path',
+        'log_handler'
     ], afset=_updatelogger
 )
 @addproperties(names=['logger'])
@@ -61,7 +84,7 @@ class Logger(Configurable):
     LOG_PATH = 'log_path'  #: logging path property name.
     LOG_DEBUG_FORMAT = 'log_debug_format'  #: debug log format property name.
     LOG_INFO_FORMAT = 'log_info_format'  #: info log format property name.
-    LOG_WARNING_FORMAT = 'log_warning_format'  #: warn log format property name
+    LOG_WARNING_FORMAT = 'log_warning_format'  #: warn log format property name.
     LOG_ERROR_FORMAT = 'log_error_format'  #: error log format property name.
     LOG_CRITICAL_FORMAT = 'log_critical_format'  #: crit log format property.
 
@@ -76,9 +99,9 @@ class Logger(Configurable):
     CRITICAL_FORMAT = ERROR_FORMAT  #: critical message format.
 
     def __init__(
-            self,
-            log_lvl='INFO', log_name=None, log_path='.',
-            log_info_format=INFO_FORMAT,
+            self, logger=None,
+            log_lvl='INFO', log_name=None, log_path=LOG_PATH,
+            log_handler=_filehandler, log_info_format=INFO_FORMAT,
             log_debug_format=DEBUG_FORMAT, log_warning_format=WARNING_FORMAT,
             log_error_format=ERROR_FORMAT, log_critical_format=CRITICAL_FORMAT,
             *args, **kwargs
@@ -88,6 +111,8 @@ class Logger(Configurable):
         :param str log_name: logger name. Default is configurable class lower
             name.
         :param str log_path: logging file path. Default is current directory.
+        :param log_handler: function which takes in parameter self configurable
+            and returns a logging Handler.
         :param str log_info_format: info logging level format.
         :param str log_debug_format: debug logging level format.
         :param str log_warning_format: warning logging level format.
@@ -97,9 +122,12 @@ class Logger(Configurable):
 
         super(Logger, self).__init__(*args, **kwargs)
 
+        self.logger = self if logger is None else logger  # self logger is self
+
         # init protected attributes
         self._log_lvl = log_lvl
         self._log_path = log_path
+        self._log_handler = log_handler
         self._log_name = log_name if log_name else type(self).__name__.lower()
         self._log_debug_format = log_debug_format
         self._log_info_format = log_info_format
@@ -109,13 +137,15 @@ class Logger(Configurable):
 
         self._logger = self.newlogger()
 
+
+
     def newlogger(self):
         """Get a new logger related to self properties."""
 
         result = getLogger(self.log_name)
         result.setLevel(self.log_lvl)
 
-        def sethandler(logger, lvl, path, _format):
+        def sethandler(logger, lvl, _format):
             """Set right handler related to input lvl, path and format.
 
             :param Logger logger: logger on which add an handler.
@@ -125,13 +155,12 @@ class Logger(Configurable):
             """
 
             class _Filter(Filter):
-                """Ensure message will be given for specific lvl
-                """
+                """Ensure message will be given for specific lvl."""
                 def filter(self, record):
                     return record.levelname == lvl
 
             # get the rights formatter and filter to set on a file handler
-            handler = FileHandler(path)
+            handler = self._log_handler(self)
             handler.addFilter(_Filter())
             handler.setLevel(lvl)
             formatter = Formatter(_format)
@@ -145,14 +174,11 @@ class Logger(Configurable):
             logger.addHandler(handler)
             setattr(logger, lvl, handler)
 
-        filename = self.log_name.replace('.', sep)
-        path = join(self.log_path, '{0}.log'.format(filename))
-
-        sethandler(result, 'DEBUG', path, self.log_debug_format)
-        sethandler(result, 'INFO', path, self.log_info_format)
-        sethandler(result, 'WARNING', path, self.log_warning_format)
-        sethandler(result, 'ERROR', path, self.log_error_format)
-        sethandler(result, 'CRITICAL', path, self.log_critical_format)
+        sethandler(result, 'DEBUG', self.log_debug_format)
+        sethandler(result, 'INFO', self.log_info_format)
+        sethandler(result, 'WARNING', self.log_warning_format)
+        sethandler(result, 'ERROR', self.log_error_format)
+        sethandler(result, 'CRITICAL', self.log_critical_format)
 
         return result
 
@@ -161,15 +187,15 @@ class Logger(Configurable):
         result = super(Logger, self)._clsconf(*args, **kwargs)
 
         result += Category(
-            Configurable.LOG,
-            Parameter(name=Configurable.LOG_NAME, critical=True),
-            Parameter(name=Configurable.LOG_PATH, critical=True),
-            Parameter(name=Configurable.LOG_LVL, critical=True),
-            Parameter(name=Configurable.LOG_DEBUG_FORMAT, critical=True),
-            Parameter(name=Configurable.LOG_INFO_FORMAT, critical=True),
-            Parameter(name=Configurable.LOG_WARNING_FORMAT, critical=True),
-            Parameter(name=Configurable.LOG_ERROR_FORMAT, critical=True),
-            Parameter(name=Configurable.LOG_CRITICAL_FORMAT, critical=True)
+            Logger.LOG,
+            Parameter(name=Logger.LOG_NAME, critical=True),
+            Parameter(name=Logger.LOG_PATH, critical=True),
+            Parameter(name=Logger.LOG_LVL, critical=True),
+            Parameter(name=Logger.LOG_DEBUG_FORMAT, critical=True),
+            Parameter(name=Logger.LOG_INFO_FORMAT, critical=True),
+            Parameter(name=Logger.LOG_WARNING_FORMAT, critical=True),
+            Parameter(name=Logger.LOG_ERROR_FORMAT, critical=True),
+            Parameter(name=Logger.LOG_CRITICAL_FORMAT, critical=True)
         )
 
         return result
@@ -195,23 +221,17 @@ class Logger(Configurable):
 
         self._logger.setLevel(self._log_lvl)
 
-    @property
-    def log_name(self):
-        """Get this logger name.
+# set logger to arguments of Logger methods
+for name, member in getmembers(Logger, isfunction):
 
-        :return: self logger name.
-        :rtype: str
-        """
+    if name != '__init__' and 'logger' in getargspec(member).args:
 
-        return self._log_name
+        @wraps(member)
+        def newmethod(self, logger=None, _name=name, *args, **kwargs):
 
-    @log_name.setter
-    def log_name(self, value):
-        """Change of logging name.
+            if logger is None:
+                logger = self.logger
 
-        :param str value: new log_name to set up.
-        """
+            return getattr(self, _name)(logger=logger, *args, **kwargs)
 
-        self._log_name = value
-
-        self._logger.setLevel(self._log_name)
+        setattr(Logger, name, newmethod)